Remove record of decision feature flag
