+++ api/config/feature.php
@@ -27,17 +27,6 @@ return [
 
     /*
     |--------------------------------------------------------------------------
-    | Record of decision
-    |--------------------------------------------------------------------------
-    |
-    | A feature to introduce a multiple step assessment builder for evaluating candidates
-    |
-    */
-
-    'record_of_decision' => (bool) env('FEATURE_RECORD_OF_DECISION', false),
-
-    /*
-    |--------------------------------------------------------------------------
     | Notifications
     |--------------------------------------------------------------------------
     |
