+++ apps/e2e/cypress/e2e/admin/pool-candidate.cy.ts
@@ -24,7 +24,6 @@ describe("Pool Candidates", () => {
   };
 
   beforeEach(() => {
-    cy.overrideFeatureFlags({ FEATURE_RECORD_OF_DECISION: false });
     cy.intercept("POST", "/graphql", (req) => {
       aliasQuery(req, "ViewPoolCandidatesPage");
       aliasQuery(req, "PoolTable");
