+++ apps/web/.env.example
@@ -18,7 +18,6 @@ OAUTH_LOGOUT_URI="http://localhost:8000/oxauth/endsession"
 #FEATURE_STATUS_NOTIFICATIONS=true  // not yet implemented on the client side
 FEATURE_NOTIFICATIONS=true
 FEATURE_DIRECTIVE_FORMS=true
-FEATURE_RECORD_OF_DECISION=true
 
 # Logging - enable by setting the logging level. Eight severity levels from rfc5424.
 LOG_CONSOLE_LEVEL="debug"
