+++ apps/web/src/pages/Applications/ApplicationQuestionsPage/ApplicationQuestionsPage.tsx
@@ -7,7 +7,6 @@ import { BasicForm } from "@gc-digital-talent/forms";
 import { notEmpty } from "@gc-digital-talent/helpers";
 import { toast } from "@gc-digital-talent/toast";
 import { ApplicationStep } from "@gc-digital-talent/graphql";
-import { useFeatureFlags } from "@gc-digital-talent/env";
 
 import useRoutes from "~/hooks/useRoutes";
 import { GetPageNavInfo } from "~/types/applicationStep";
@@ -27,13 +26,10 @@ export const getPageInfo: GetPageNavInfo = ({
   paths,
   intl,
   stepOrdinal,
-  RoDFlag,
 }) => {
   const path = paths.applicationQuestions(application.id);
   return {
-    title: RoDFlag
-      ? intl.formatMessage(processMessages.additionalQuestions)
-      : intl.formatMessage(processMessages.screeningQuestions),
+    title: intl.formatMessage(processMessages.additionalQuestions),
     subtitle: intl.formatMessage({
       defaultMessage: "Answer key questions about your fit in this role.",
       id: "GTHuSJ",
@@ -56,16 +52,14 @@ export const getPageInfo: GetPageNavInfo = ({
 const ApplicationQuestions = ({ application }: ApplicationPageProps) => {
   const intl = useIntl();
   const paths = useRoutes();
-  const features = useFeatureFlags();
   const navigate = useNavigate();
   const { isIAP } = useApplicationContext();
   const [{ fetching: mutating }, executeMutation] =
     useUpdateApplicationMutation();
   const cancelPath = paths.profileAndApplications({ fromIapDraft: isIAP });
 
-  const screeningQuestions = features.recordOfDecision
-    ? application.pool.screeningQuestions?.filter(notEmpty) || []
-    : [];
+  const screeningQuestions =
+    application.pool.screeningQuestions?.filter(notEmpty) || [];
   const screeningQuestionResponses =
     application.screeningQuestionResponses?.filter(notEmpty) || [];
   const generalQuestions =
@@ -137,7 +131,7 @@ const ApplicationQuestions = ({ application }: ApplicationPageProps) => {
       }}
     >
       {/* Screening Questions */}
-      {screeningQuestions.length > 0 && features.recordOfDecision && (
+      {screeningQuestions.length > 0 && (
         <div data-h2-margin-bottom="base(x4)">
           <div
             data-h2-display="p-tablet(flex)"
@@ -199,7 +193,6 @@ const ApplicationQuestions = ({ application }: ApplicationPageProps) => {
           ))}
         </div>
       )}
-
       {/* General Questions */}
       {generalQuestions.length > 0 && (
         <>
@@ -216,8 +209,7 @@ const ApplicationQuestions = ({ application }: ApplicationPageProps) => {
             >
               {intl.formatMessage(processMessages.generalQuestions)}
             </Heading>
-            {(!features.recordOfDecision ||
-              screeningQuestions.length === 0) && (
+            {screeningQuestions.length === 0 && (
               <Link
                 color="secondary"
                 mode="inline"
