+++ apps/web/src/pages/Applications/ApplicationReviewPage/ApplicationReviewPage.tsx
@@ -16,7 +16,6 @@ import { errorMessages, getLocale } from "@gc-digital-talent/i18n";
 import { Input } from "@gc-digital-talent/forms";
 import { toast } from "@gc-digital-talent/toast";
 import { Experience, SkillCategory, graphql } from "@gc-digital-talent/graphql";
-import { useFeatureFlags } from "@gc-digital-talent/env";
 
 import useRoutes from "~/hooks/useRoutes";
 import { GetPageNavInfo } from "~/types/applicationStep";
@@ -93,7 +92,6 @@ const ApplicationReview = ({
   const intl = useIntl();
   const locale = getLocale(intl);
   const paths = useRoutes();
-  const features = useFeatureFlags();
   const navigate = useNavigate();
   const { currentStepOrdinal, followingPageUrl, isIAP } =
     useApplicationContext();
@@ -102,7 +100,6 @@ const ApplicationReview = ({
     paths,
     application,
     stepOrdinal: currentStepOrdinal,
-    RoDFlag: features.recordOfDecision,
   });
   const nextStep = followingPageUrl ?? paths.applicationSuccess(application.id);
 
@@ -406,169 +403,111 @@ const ApplicationReview = ({
           )}
         </div>
       </ReviewSection>
-      {features.recordOfDecision ? (
-        <>
-          {screeningQuestions.length > 0 && (
-            <ReviewSection
-              title={intl.formatMessage(processMessages.screeningQuestions)}
-              path={editPaths.applicationQuestions}
-              editLinkAriaLabel={intl.formatMessage({
-                defaultMessage: "Edit screening questions",
-                id: "5A0a7w",
-                description:
-                  "Edit link text for screening questions section of the application review page.",
-              })}
-            >
-              {screeningQuestionResponses.length > 0 ? (
-                <div>
-                  <p data-h2-margin="base(x1, 0, x.5, 0)">
-                    {intl.formatMessage({
-                      defaultMessage:
-                        "You’ve answered the following screening questions:",
-                      id: "Sd8883",
-                      description:
-                        "Message in screening questions section of the application review page.",
-                    })}
-                  </p>
-                  <ul data-h2-padding="base(0, 0, 0, x1)">
-                    {screeningQuestionResponses.map((response) => (
-                      <li key={response.id} data-h2-margin-bottom="base(x.5)">
-                        <p
-                          data-h2-font-weight="base(700)"
-                          data-h2-margin-bottom="base(x.25)"
-                        >
-                          {response.screeningQuestion?.question
-                            ? response.screeningQuestion.question[locale]
-                            : ""}
-                        </p>
-                        <p>{response.answer}</p>
-                      </li>
-                    ))}
-                  </ul>
-                </div>
-              ) : (
-                <Well>
-                  <p data-h2-text-align="base(center)">
-                    {intl.formatMessage({
-                      defaultMessage:
-                        "It looks like you haven't answered any screening questions yet.",
-                      id: "V9lxDE",
-                      description:
-                        "Null state message in screening questions section of the application review page.",
-                    })}
-                  </p>
-                </Well>
-              )}
-            </ReviewSection>
+      {screeningQuestions.length > 0 && (
+        <ReviewSection
+          title={intl.formatMessage(processMessages.screeningQuestions)}
+          path={editPaths.applicationQuestions}
+          editLinkAriaLabel={intl.formatMessage({
+            defaultMessage: "Edit screening questions",
+            id: "5A0a7w",
+            description:
+              "Edit link text for screening questions section of the application review page.",
+          })}
+        >
+          {screeningQuestionResponses.length > 0 ? (
+            <div>
+              <p data-h2-margin="base(x1, 0, x.5, 0)">
+                {intl.formatMessage({
+                  defaultMessage:
+                    "You’ve answered the following screening questions:",
+                  id: "Sd8883",
+                  description:
+                    "Message in screening questions section of the application review page.",
+                })}
+              </p>
+              <ul data-h2-padding="base(0, 0, 0, x1)">
+                {screeningQuestionResponses.map((response) => (
+                  <li key={response.id} data-h2-margin-bottom="base(x.5)">
+                    <p
+                      data-h2-font-weight="base(700)"
+                      data-h2-margin-bottom="base(x.25)"
+                    >
+                      {response.screeningQuestion?.question
+                        ? response.screeningQuestion.question[locale]
+                        : ""}
+                    </p>
+                    <p>{response.answer}</p>
+                  </li>
+                ))}
+              </ul>
+            </div>
+          ) : (
+            <Well>
+              <p data-h2-text-align="base(center)">
+                {intl.formatMessage({
+                  defaultMessage:
+                    "It looks like you haven't answered any screening questions yet.",
+                  id: "V9lxDE",
+                  description:
+                    "Null state message in screening questions section of the application review page.",
+                })}
+              </p>
+            </Well>
           )}
-          {generalQuestions.length > 0 && (
-            <ReviewSection
-              title={intl.formatMessage(processMessages.generalQuestions)}
-              path={editPaths.applicationQuestions}
-              editLinkAriaLabel={intl.formatMessage({
-                defaultMessage: "Edit general questions",
-                id: "gIzUDr",
-                description:
-                  "Edit link text for general questions section of the application review page.",
-              })}
-            >
-              {generalQuestionResponses.length > 0 ? (
-                <div>
-                  <p data-h2-margin="base(x1, 0, x.5, 0)">
-                    {intl.formatMessage({
-                      defaultMessage:
-                        "You've answered the following general questions:",
-                      id: "96AvU9",
-                      description:
-                        "Message in general questions section of the application review page.",
-                    })}
-                  </p>
-                  <ul data-h2-padding="base(0, 0, 0, x1)">
-                    {generalQuestionResponses.map((response) => (
-                      <li key={response.id} data-h2-margin-bottom="base(x.5)">
-                        <p
-                          data-h2-font-weight="base(700)"
-                          data-h2-margin-bottom="base(x.25)"
-                        >
-                          {response.generalQuestion?.question
-                            ? response.generalQuestion.question[locale]
-                            : ""}
-                        </p>
-                        <p>{response.answer}</p>
-                      </li>
-                    ))}
-                  </ul>
-                </div>
-              ) : (
-                <Well>
-                  <p data-h2-text-align="base(center)">
-                    {intl.formatMessage({
-                      defaultMessage:
-                        "It looks like you haven't answered any general questions yet.",
-                      id: "OPbfwn",
-                      description:
-                        "Null state message in general questions section of the application review page.",
-                    })}
-                  </p>
-                </Well>
-              )}
-            </ReviewSection>
+        </ReviewSection>
+      )}
+      {generalQuestions.length > 0 && (
+        <ReviewSection
+          title={intl.formatMessage(processMessages.generalQuestions)}
+          path={editPaths.applicationQuestions}
+          editLinkAriaLabel={intl.formatMessage({
+            defaultMessage: "Edit general questions",
+            id: "gIzUDr",
+            description:
+              "Edit link text for general questions section of the application review page.",
+          })}
+        >
+          {generalQuestionResponses.length > 0 ? (
+            <div>
+              <p data-h2-margin="base(x1, 0, x.5, 0)">
+                {intl.formatMessage({
+                  defaultMessage:
+                    "You've answered the following general questions:",
+                  id: "96AvU9",
+                  description:
+                    "Message in general questions section of the application review page.",
+                })}
+              </p>
+              <ul data-h2-padding="base(0, 0, 0, x1)">
+                {generalQuestionResponses.map((response) => (
+                  <li key={response.id} data-h2-margin-bottom="base(x.5)">
+                    <p
+                      data-h2-font-weight="base(700)"
+                      data-h2-margin-bottom="base(x.25)"
+                    >
+                      {response.generalQuestion?.question
+                        ? response.generalQuestion.question[locale]
+                        : ""}
+                    </p>
+                    <p>{response.answer}</p>
+                  </li>
+                ))}
+              </ul>
+            </div>
+          ) : (
+            <Well>
+              <p data-h2-text-align="base(center)">
+                {intl.formatMessage({
+                  defaultMessage:
+                    "It looks like you haven't answered any general questions yet.",
+                  id: "OPbfwn",
+                  description:
+                    "Null state message in general questions section of the application review page.",
+                })}
+              </p>
+            </Well>
           )}
-        </>
-      ) : (
-        generalQuestions.length > 0 && (
-          <ReviewSection
-            title={intl.formatMessage(processMessages.screeningQuestions)}
-            path={editPaths.applicationQuestions}
-            editLinkAriaLabel={intl.formatMessage({
-              defaultMessage: "Edit screening questions",
-              id: "5A0a7w",
-              description:
-                "Edit link text for screening questions section of the application review page.",
-            })}
-          >
-            {generalQuestionResponses.length > 0 ? (
-              <div>
-                <p data-h2-margin="base(x1, 0, x.5, 0)">
-                  {intl.formatMessage({
-                    defaultMessage:
-                      "You’ve answered the following screening questions:",
-                    id: "Sd8883",
-                    description:
-                      "Message in screening questions section of the application review page.",
-                  })}
-                </p>
-                <ul data-h2-padding="base(0, 0, 0, x1)">
-                  {generalQuestionResponses.map((response) => (
-                    <li key={response.id} data-h2-margin-bottom="base(x.5)">
-                      <p
-                        data-h2-font-weight="base(700)"
-                        data-h2-margin-bottom="base(x.25)"
-                      >
-                        {response.generalQuestion?.question
-                          ? response.generalQuestion.question[locale]
-                          : ""}
-                      </p>
-                      <p>{response.answer}</p>
-                    </li>
-                  ))}
-                </ul>
-              </div>
-            ) : (
-              <Well>
-                <p data-h2-text-align="base(center)">
-                  {intl.formatMessage({
-                    defaultMessage:
-                      "It looks like you haven't answered any screening questions yet.",
-                    id: "V9lxDE",
-                    description:
-                      "Null state message in screening questions section of the application review page.",
-                  })}
-                </p>
-              </Well>
-            )}
-          </ReviewSection>
-        )
+        </ReviewSection>
       )}
 
       <section data-h2-margin="base(x3, 0, 0, 0)">
