+++ apps/web/src/pages/Applications/questionsStep/questionsStepValidation.tsx
@@ -9,11 +9,10 @@ const stepHasError = (
   _user: User,
   pool: Pool,
   application: Omit<PoolCandidate, "pool">,
-  RoDFlag: boolean,
 ) => {
   return (
     generalQuestionsSectionHasMissingResponses(application, pool) ||
-    screeningQuestionsSectionHasMissingResponses(application, pool, RoDFlag)
+    screeningQuestionsSectionHasMissingResponses(application, pool)
   );
 };
