+++ apps/web/src/pages/Applications/selfDeclarationStep/selfDeclarationStepInfo.tsx
@@ -13,7 +13,6 @@ const getStepInfo: GetApplicationStepInfo = ({
   paths,
   intl,
   stepOrdinal,
-  RoDFlag,
 }): ApplicationStepInfo => {
   return {
     applicationStep: ApplicationStep.SelfDeclaration,
@@ -22,7 +21,6 @@ const getStepInfo: GetApplicationStepInfo = ({
       intl,
       application,
       stepOrdinal,
-      RoDFlag,
     }),
     showInStepper: true,
     prerequisites: [ApplicationStep.Welcome],
