+++ apps/web/src/pages/Applications/successStep/successStepInfo.tsx
@@ -12,7 +12,6 @@ const getStepInfo: GetApplicationStepInfo = ({
   paths,
   intl,
   stepOrdinal,
-  RoDFlag,
 }): ApplicationStepInfo => {
   return {
     mainPage: successPageInfo({
@@ -20,7 +19,6 @@ const getStepInfo: GetApplicationStepInfo = ({
       intl,
       application,
       stepOrdinal,
-      RoDFlag,
     }),
     showInStepper: false,
     prerequisites: [
