+++ apps/web/src/pages/PoolCandidates/ViewPoolCandidatePage/components/CareerTimelineSection/CareerTimelineSection.tsx
@@ -1,10 +1,8 @@
 import * as React from "react";
 import { useIntl } from "react-intl";
-import UserCircleIcon from "@heroicons/react/24/outline/UserCircleIcon";
 
 import { notEmpty } from "@gc-digital-talent/helpers";
-import { Heading, Well } from "@gc-digital-talent/ui";
-import { useFeatureFlags } from "@gc-digital-talent/env";
+import { Well } from "@gc-digital-talent/ui";
 import { Experience } from "@gc-digital-talent/graphql";
 
 import ExperienceCard from "~/components/ExperienceCard/ExperienceCard";
@@ -18,7 +16,6 @@ interface CareerTimelineSectionProps {
 }
 
 const CareerTimelineSection = ({ experiences }: CareerTimelineSectionProps) => {
-  const features = useFeatureFlags();
   const intl = useIntl();
   const [sortAndFilterValues, setSortAndFilterValues] =
     React.useState<ExperienceSortAndFilterFormValues>({
@@ -36,15 +33,6 @@ const CareerTimelineSection = ({ experiences }: CareerTimelineSectionProps) => {
 
   return (
     <>
-      {features.recordOfDecision && (
-        <Heading Icon={UserCircleIcon} color="tertiary">
-          {intl.formatMessage({
-            defaultMessage: "Career timeline",
-            id: "5FK/4R",
-            description: "Heading for the users experiences",
-          })}
-        </Heading>
-      )}
       <div
         data-h2-flex-grid="base(center, x1, x1)"
         data-h2-margin-bottom="base(x.5)"
