+++ apps/web/src/pages/Pools/ViewPoolPage/ViewPoolPage.tsx
@@ -12,7 +12,6 @@ import {
   parseDateTimeUtc,
 } from "@gc-digital-talent/date-helpers";
 import { ROLE_NAME, useAuthorization } from "@gc-digital-talent/auth";
-import { useFeatureFlags } from "@gc-digital-talent/env";
 import { graphql, Pool, PoolStatus, Scalars } from "@gc-digital-talent/graphql";
 
 import SEO from "~/components/SEO/SEO";
@@ -73,7 +72,6 @@ export const ViewPool = ({
     [ROLE_NAME.CommunityManager, ROLE_NAME.PlatformAdmin],
     roleAssignments,
   );
-  const { recordOfDecision: recordOfDecisionFlag } = useFeatureFlags();
 
   let closingDate = "";
   if (pool.closingDate) {
@@ -103,10 +101,9 @@ export const ViewPool = ({
     description: "Subtitle for the individual pool page",
   });
 
-  const isReadyToPublish = recordOfDecisionFlag
-    ? getAdvertisementStatus(pool) === "complete" &&
-      getAssessmentPlanStatus(pool) === "complete"
-    : getAdvertisementStatus(pool) === "complete";
+  const isReadyToPublish =
+    getAdvertisementStatus(pool) === "complete" &&
+    getAssessmentPlanStatus(pool) === "complete";
 
   return (
     <>
@@ -197,14 +194,9 @@ export const ViewPool = ({
               <Heading level="h3" size="h6" data-h2-margin="base(0)">
                 {intl.formatMessage(messages.assessmentPlan)}
               </Heading>
-              {recordOfDecisionFlag && (
-                <Chip
-                  color={assessmentBadge.color}
-                  data-h2-flex-shrink="base(0)"
-                >
-                  {intl.formatMessage(assessmentBadge.label)}
-                </Chip>
-              )}
+              <Chip color={assessmentBadge.color} data-h2-flex-shrink="base(0)">
+                {intl.formatMessage(assessmentBadge.label)}
+              </Chip>
             </ProcessCard.Header>
             <p data-h2-margin="base(x1 0)">
               {intl.formatMessage({
@@ -216,34 +208,25 @@ export const ViewPool = ({
               })}
             </p>
             <ProcessCard.Footer>
-              {recordOfDecisionFlag ? (
-                <Link
-                  mode="inline"
-                  color="secondary"
-                  href={paths.assessmentPlanBuilder(pool.id)}
-                >
-                  {assessmentStatus === "submitted"
-                    ? intl.formatMessage({
-                        defaultMessage: "View assessment plan",
-                        id: "1X7JVN",
-                        description:
-                          "Link text to view a specific pool assessment",
-                      })
-                    : intl.formatMessage({
-                        defaultMessage: "Edit assessment plan",
-                        id: "Q3adCp",
-                        description:
-                          "Link text to edit a specific pool assessment",
-                      })}
-                </Link>
-              ) : (
-                intl.formatMessage({
-                  defaultMessage: "Coming soon",
-                  id: "/IMv2G",
-                  description:
-                    "Message displayed when a feature is in development and not ready yet",
-                })
-              )}
+              <Link
+                mode="inline"
+                color="secondary"
+                href={paths.assessmentPlanBuilder(pool.id)}
+              >
+                {assessmentStatus === "submitted"
+                  ? intl.formatMessage({
+                      defaultMessage: "View assessment plan",
+                      id: "1X7JVN",
+                      description:
+                        "Link text to view a specific pool assessment",
+                    })
+                  : intl.formatMessage({
+                      defaultMessage: "Edit assessment plan",
+                      id: "Q3adCp",
+                      description:
+                        "Link text to edit a specific pool assessment",
+                    })}
+              </Link>
             </ProcessCard.Footer>
           </ProcessCard.Root>
           <ProcessCard.Root data-h2-grid-column="l-tablet(span 2)">
