+++ apps/web/src/utils/applicationUtils.ts
@@ -29,7 +29,6 @@ type GetApplicationPagesArgs = {
   intl: IntlShape;
   application: Application_PoolCandidateFragment;
   experienceId?: string;
-  RoDFlag: boolean;
 };
 
 // Dynamically build the list of application steps for this application
@@ -38,11 +37,10 @@ export const getApplicationSteps = ({
   intl,
   application,
   experienceId,
-  RoDFlag,
 }: GetApplicationPagesArgs): Array<ApplicationStepInfo> => {
   const showQuestionStep =
     application.pool.generalQuestions?.length ||
-    (RoDFlag && application.pool.screeningQuestions?.length);
+    application.pool.screeningQuestions?.length;
 
   // build the order of step functions to call
   const stepInfoFunctions = [
@@ -65,7 +63,6 @@ export const getApplicationSteps = ({
       application,
       resourceId: experienceId,
       stepOrdinal: index + 1,
-      RoDFlag,
     }),
   );
 
@@ -146,7 +143,6 @@ export function isOnDisabledPage(
 export function applicationStepsToStepperArgs(
   applicationSteps: Array<ApplicationStepInfo>,
   application: PoolCandidate,
-  RoDFlag: boolean,
 ): StepType[] {
   return applicationSteps
     .filter((step) => step.showInStepper)
@@ -162,12 +158,7 @@ export function applicationStepsToStepperArgs(
           step.prerequisites,
           application.submittedSteps,
         )?.length,
-        error: step.hasError?.(
-          application.user,
-          application.pool,
-          application,
-          RoDFlag,
-        ),
+        error: step.hasError?.(application.user, application.pool, application),
       };
     });
 }
